Enforce that the JDBC meta column count matches the amount of data we have.
An assert is sufficient, here, as this is a library bug rather than a program issue, and there really is no recovery from it anyhow. Assertions are enabled in our tests now, and most developers will run tests with assertions enabled (default behavior for Maven and Gradle).

This would have more quickly identified #6301