Java: Faster DuckDBResultSetMetaData creation
Reduce the amount of work happening in DuckDBResultSetMetaData. This class is created for every ResultSet, so changes here have a direct impact on callers. Overall, the cost of creation is tiny compared to JNI, IO, and query, but these changes are neither convoluted nor difficult.

Rather than converting Arrays to Lists and back to Arrays, use the Stream API to map the Array into a List, and keep it as a list. The JVM is able to optimize the map() operation, and index lookups on the list are comparable to those on Arrays.

Change the string replacement logic to use a pre-compiled Regex to avoid regex compilation, which can be notoriously slow on the JVM.

Stop storing column_type_details as it is unused, and we don't need the memory overhead.