Java: Move DuckDBResultSetMetaData#type_to_int mapping to DuckDBColumnType
Java enums are full classes, and so can have fields, constructors, and methods. Mapping from an enum type to an integer, basically, associating the enum with a value, is an idea use case for the enum to have a field. The benefit of this is faster lookups and it is impossible to add a type without also mapping it correctly.

The tests that were verifying metadata, now do nothing more than verify configuration, so those lines have been removed.